RELEASE 2019-03-18 (API v3.4.1)

+++ src/main/java/com/neutrinoapi/sdk/models/BrowserBotResponseBuilder.java
@@ -125,7 +125,7 @@ public class BrowserBotResponseBuilder {
     /**
      * The number of seconds taken to load the page (from initial request until DOM ready)
      */
-    public BrowserBotResponseBuilder loadTime(int loadTime) {
+    public BrowserBotResponseBuilder loadTime(double loadTime) {
         browserBotResponse.setLoadTime(loadTime);
         return this;
     }
